// hld/sdk/typescript/src/generated/runtime.ts
export type FetchAPI = WindowOrWorkerGlobalScope['fetch'];
// ...
export interface ErrorContext {
    fetch: FetchAPI;
    url: string;
    init: RequestInit;
    error: unknown;
    response?: Response;
}